+++ bin/cli.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+import { default as main } from "../out/src/cli/index.js";
+main();