+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 - Updated to Langium 2.0.0
 - Updated monaco-editor-wrapper to the latest version (3.1.0)
 - Updated the built-in web-app to use the latest wrapper & langium versions, plus some freshening up
+- Reworked the CLI to directly output generated Minilogo drawing commands, so its output matches that of the document listener on the language server
+- Reworked the old generate:test script (and dropped generate:logo) to just output the CLI's generated drawing commands to the console
 
 # 1.2.0
 
 - Bumped to Langium version 1.2.0
 - Updated associated packages to support usage of the new [langium-minilogo npm package](https://www.npmjs.com/package/langium-minilogo)
+- Setup a document listener in the language server, allowing the client to register for & receive Minilogo commands from validated documents
 
 # 1.0.0